HW 1. Small refactoring
