+++ src/test/java/hw1/CalculatorBaseTest.java
@@ -1,20 +1,21 @@
-package hw1.arithmetic;
+package hw1;
 
 import com.epam.tat.module4.Calculator;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeMethod;
+import org.testng.annotations.AfterClass;
+import org.testng.annotations.BeforeClass;
 
 public class CalculatorBaseTest {
 
     protected Calculator calculator;
     protected static final double DOUBLE_DELTA = 1e-6;
 
-    @BeforeMethod
+    @BeforeClass
     public void setUp() {
+        System.out.println("213");
         calculator = new Calculator();
     }
 
-    @AfterMethod
+    @AfterClass
     public void tearDown() {
         calculator = null;
     }
