+++ src/test/java/hw1/arithmetic/MultOperationTest.java
@@ -1,5 +1,6 @@
 package hw1.arithmetic;
 
+import hw1.CalculatorBaseTest;
 import hw1.data_providers.ArithmeticDataProviders;
 import org.testng.annotations.Test;
